Test creating bootable disk

+++ bin/infra-vm-image.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
 
-# $1 - efi
-# $2 - rootfs
+# $1 - efi directory if exists
+# $2 - rootfs directory if exists
+
+# $3 - target : all (default), vm, rootfs
+
+if [ -z "$3" ]; then
+  TARGET="all"
+else
+  TARGET="$3"
+fi
 
 FILENAME=linux-flat.vmdk
 OUT_DIR=${OUT_DIR:=/tmp/img}
@@ -13,7 +21,7 @@ MNT_EFI=$MNT_DIR/efi
   mkdir -p $MNT $MNT_EFI
   echo "Installing $RELEASE into $FILE..."
 
-  # 7GB image file to fit comfortable to 8 GB sticks or larger
+  # 3GB image file to fit comfortable to 8 GB sticks or larger
   IMGSIZE=${IMGSIZE:=3072} # in megabytes
 
   if [ -f $FILE ]; then
@@ -46,20 +54,21 @@ MNT_EFI=$MNT_DIR/efi
   # See also https://systemd.io/DISCOVERABLE_PARTITIONS/
   # https://wiki.archlinux.org/title/GPT_fdisk
   sudo sgdisk -Z $DISK
-  sudo sgdisk -n 0:0:+510M  -t 0:ef00 -c 0:"efi_vm"   $DISK
+
+  # Add efi partition
+  if [ "$TARGET" != "rootfs" ]; then
+    sudo sgdisk -n 0:0:+510M  -t 0:ef00 -c 0:"efi_vm" $DISK
+    sudo mkfs.vfat -F 32 -n EFI -i 10000000 ${DISK}p1 || fail "cannot create efi"
+    sudo mount ${DISK}p1 $MNT_EFI || fail "cannot mount"
+  fi
+
   sudo sgdisk -n 0:0:       -t 0:8304 -c 0:"linux_vm" $DISK
   sudo partprobe $DISK
 
   echo "Creating filesystems..."
-  sudo mkfs.vfat -F 32 -n EFI -i 10000000 ${DISK}p1 || fail "cannot create efi"
   sudo mkfs.btrfs -L "linux" -U 10000000-0000-0000-0000-000000000000 ${DISK}p2 || fail "cannot create root"
 
-  # Mount device
-  echo "Mounting partitions..."
-  sudo mount ${DISK}p1 $MNT_EFI || fail "cannot mount"
-
   # Todo - test compress
-  # -o compress
   sudo mount -o compress ${DISK}p2 $MNT || fail "cannot mount"
   sudo chmod g+w  $MNT
   sudo btrfs subvolume create $MNT/linux
@@ -83,7 +92,7 @@ if [ -z $1 ]; then
   infra-get-efi.sh
   sudo rsync -r /tmp/efi/efi/ $MNT_EFI
 else
-  sudo rsync -av $1 $MNT/linux/
+  sudo rsync -av $1 $MNT_EFI
 fi
 
   # https://wiki.archlinux.org/title/Syslinux
@@ -93,7 +102,10 @@ fi
   cd /
 
   # directories that are not needed for vm
-  sudo rm -rf $MNT_EFI/syslinux $MNT_EFI/tce
+  if [ "$TARGET" == vm ]; then
+    sudo rm -rf $MNT_EFI/syslinux $MNT_EFI/tce
+  fi
+
   sudo umount $MNT_EFI
 
   sudo losetup -d /dev/loop* 2>/dev/null
